impls copy apis for DecryptedRead
