+++ src/relay/tcprelay/server.rs
@@ -32,7 +32,7 @@ use config::{Config, ServerConfig};
 use relay::socks5::Address;
 use relay::{BoxIoFuture, boxed_future};
 use relay::dns_resolver::DnsResolver;
-use relay::tcprelay::crypto_io::EncryptedWrite;
+use relay::tcprelay::crypto_io::{EncryptedWrite, DecryptedRead};
 
 use futures::{self, Future};
 use futures::stream::Stream;
@@ -45,7 +45,7 @@ use net2::TcpBuilder;
 
 use ip::IpAddr;
 
-use super::{tunnel, proxy_handshake, DecryptedHalf, EncryptedHalfFut, try_timeout, copy_timeout};
+use super::{tunnel, proxy_handshake, DecryptedHalf, EncryptedHalfFut, try_timeout};
 
 /// Context for doing handshake with client
 pub struct TcpRelayClientHandshake {
@@ -187,7 +187,7 @@ impl TcpRelayClientConnected {
         let timeout = self.timeout;
         let handle = self.handle.clone();
         tunnel(self.addr,
-               copy_timeout(r, svr_w, self.timeout, self.handle),
+               r.copy_timeout_opt(svr_w, self.timeout, self.handle),
                w_fut.and_then(move |w| w.copy_timeout_opt(svr_r, timeout, handle)))
     }
 }
